+++ docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### [Unreleased]
+- Added ability to detect podcast content
+- New database field `has_podcast` to track podcast content
 - Added ability to detect infographics in content
 - New database field `has_infographic` to track content with infographics
 - Added ability to detect videos in social media posts